// packages/rath-client/src/interfaces.ts
export enum RealTimeMode {
    NONE = 'none',
    POLLING_REPLACE = 'polling_replace',
    POLLING_APPEND = 'polling_append',
}

export enum RealTimeStatus {
    IDLE = 'idle',
    RUNNING = 'running',
    PAUSED = 'paused',
    ERROR = 'error',
}